beautifulSoap_projects/fetch_from_amazon/requirements.txt
beautifulsoup4==4.11.2
requests==2.28.1
python-dotenv==1.0.0